+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Gsea plot now loads preranked data.
 - Fixed p-value and threshold filter on volcano by adding proper bounds and removing unneccessary log conversion.
 - Bug fixes and updates.
 - Revised save table dialog.